test: move jest snapshot serializer into separate file for future reusablity
